requirements.txt
Pillow==10.0.0
PyQt5==5.15.7
requests==2.32.3
duckduckgo_search==8.0.1
cx-freeze==8.3.0
